convert prices to bigints, begin unit testing

+++ src/main/java/finance/dto/StockDTO.java
@@ -1,4 +1,6 @@
 package finance.dto;
 
-public record StockDTO(String companyName, String symbol, Double latestPrice) {
+import java.math.BigDecimal;
+
+public record StockDTO(String companyName, String symbol, BigDecimal latestPrice) {
 }
