+++ src/main/java/finance/dto/UserDTO.java
@@ -2,7 +2,9 @@ package finance.dto;
 
 import finance.entity.User;
 
-public record UserDTO(String username, String email, String firstName, String lastName, Double balance) {
+import java.math.BigDecimal;
+
+public record UserDTO(String username, String email, String firstName, String lastName, BigDecimal balance) {
     public UserDTO(User user) {
         this(user.getUsername(), user.getEmail(), user.getFirstName(), user.getLastName(), user.getBalance());
     }
