+++ src/main/java/finance/entity/Holding.java
@@ -3,18 +3,12 @@ package finance.entity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 
-import java.time.LocalDateTime;
-
-import static finance.entity.TransactionType.BUY;
-import static finance.entity.TransactionType.SELL;
-
 @Entity
 @Table(name = "holdings")
 public class Holding {
 
     @EmbeddedId
     private HoldingId id;
-
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false, updatable = false)
@@ -27,12 +21,12 @@ public class Holding {
 
     @NotNull
     @Column(nullable = false)
-    private Integer shares;
+    private Long shares;
 
     public Holding() {
     }
 
-    public Holding(User user, String symbol, String companyName, Integer shares) {
+    public Holding(User user, String symbol, String companyName, Long shares) {
         this();
         this.user = user;
         this.id = new HoldingId(user.getId(), symbol);
@@ -40,19 +34,19 @@ public class Holding {
         this.shares = shares;
     }
 
-    public void remove(int quantity) {
+    public void remove(Long quantity) {
         if (quantity <= 0) throw new IllegalArgumentException("Transaction must be a positive number of shares.");
         if (quantity > shares)
             throw new IllegalArgumentException("Insufficient shares to sell.");
         this.shares -= quantity;
     }
 
-    public void add(int quantity) {
+    public void add(Long quantity) {
         if (quantity <= 0) throw new IllegalArgumentException("Transaction must be a positive number of shares.");
         this.shares += quantity;
     }
 
-    public Integer getShares() {
+    public Long getShares() {
         return this.shares;
     }
